Fix bind --overwrite failing on first run (#4569)

+++ cli/src/cmd/forge/bind.rs
@@ -221,7 +221,7 @@ impl Cmd for BindArgs {
             return self.check_existing_bindings(&artifacts)
         }
 
-        if self.overwrite {
+        if self.overwrite && self.bindings_exist(&artifacts) {
             fs::remove_dir_all(self.bindings_root(&artifacts))?;
         }
